【修改描述】修改underfs中listall 【DTS/CR/AR单号】#176 【合入版本】BeiMing-K 24.0.0,
Match-id-f98ff9e0d3d8100abb3a516286afdba9e5994a81

+++ src/cache/cache_def.h
@@ -39,7 +39,7 @@ struct CacheAttr {
 };
 
 struct CacheObjStat {
-    uint32_t size; // value size
+    uint64_t size; // value size
     time_t   time; // modify time;
 };
 }
